ENH: fix MetaIO in VTK
MetaIO: headers were installed in /include/vtk-5.2/Utilities/MetaIO for VTK
and in /include/InsightToolkit/Utilities/MetaIO for ITK. While this seems OK
for ITK, this not for VTK; the support libs found in VTK's Utilities install
their headers in /include/vtk-5.2 (the extra 'Utilities' subdir would prevent
them from being found). The metaIOConfig.h header was not installed.

Furthermore, headers were installed in "/MetaIO" instead of "/vtkmetaio",
which seems to be the required syntax (as implemented in a few VTK's
classes: #include "vtkmetaio/metaTypes.h"). MetaIO is also missing support
for a system's MetaIO vs. a VTK's MetaIO (as demonstrated by VTK's expat,
freetype, jpeg, tiff, etc): I'll leave that to MetaIO people.

Finally, VTK_USE_METAIO was not exported at all: since this is an option,
people could link against an installed VTK and expect MetaIO, whereas it was
never built.... Considering that MetaIO is not as widely used as PNG, or
JPEG, I didn't tackle the support for a system's MetaIO library.

SVN-Revision: 198