BUG: MetaImage data file was missing path.
When writing ROIs the filename of the data was missing the path that
should make it be stored in the same directory as the header file.

Also unified the hard-coded size of buffers used to store path names.
They are now using a common MAXPATHLENGHT defined symbol set to 2048.

Change-Id: If6922e5676df8539bffe894cab514e08f7ade9e3

SVN-Revision: 3548