BUG: Missing one case in MET_SetFileSuffix
When adding a suffix to a file name in the MetaIO, the function
MET_SetFileSuffix is used. It checks if the file name already has
an extension, and in that case removes the extension. If it has an
extension, and the new extension starts with '.', it makes sure to
keep only one of the two '.'.
However, if the file name has no extension, and there no was check
to verify that the given extension started with '.'. This lead to
create file names without a '.' and with the file name concatenated
with the extension with no separator. This patch fixes this issue.