BUG: ObjectTypeName must be set in Clear() of derived types.
Previously ObjectTypeName was only set for metaObject (top level) in
the Clear() function and was set in the M_SetupWriteFields() of
derived classes.  Therefore, the value was fixed at "Object" until
the write command was called or a file was read (that then set the
type).

In this fixed version, the ObjectTypeName value is set in the Clear()
function of metaObject and its derived classes (and not in the
M_SetupWriteFields() function).  Therefore, its value is correct
as soon as the class is instantiated.