STYLE: Enforce consistent style with clang-format
```bash
find . -name "*.h" -o -name "*.[thc]xx" \
  -exec ~/local/clang-format-8.0.0-macosx -i {} \;
```

Most of the changes make inconsistent coding practices consistent
throughout the toolkit.  Community discussions regarding changes
to the current defined coding style are explicitly not included
in this request so that they can be independantly reviewed and
considered.

The coding style was chosen to match the ITK coding conventions:
Many ITK developers provided insightful comments and suggestions that
resulted in this set of style comprimises.