STYLE: Remove unused include directives
The changes include directives that are not used or not
required and can be safely removed.

Detect completely unused an include directive is unused if declarations
from it are not used in the whole translation unit at all.  Detect not
directly used an include directive is unused if declarations from it are
not used in the file directly.