Fixed cppcheck variableScope warnings
Specifically:

src/metaFEMObject.cxx:308,style,The scope of the variable 'clID' can be reduced.
src/metaForm.cxx:488,style,The scope of the variable 'out' can be reduced.
src/metaObject.cxx:1764,style,The scope of the variable 'out' can be reduced.
src/metaUtils.cxx:1019,style,The scope of the variable 'i' can be reduced.
src/metaUtils.cxx:1026,style,The scope of the variable 'found' can be reduced.
src/metaUtils.cxx:1028,style,The scope of the variable 'c' can be reduced.
src/metaUtils.cxx:835,style,The scope of the variable 'j' can be reduced.