ENH: added new variant of MET_DoubleToValue() with buffer size
MET_DoubleToValue() didn't know the size of the `_data` buffer, and so it was impossible to implement it safely, that is, with guaranteeing the buffer is not overwritten.

So , added a new variant where the buffer size is passed. Changed many, but not all, callers to use the new variant. For some callers, it wasn't clear what the buffer size is.

This allowed replacing one of the last sprintf in the codebase with snprintf, now that we can know the buffer size.