ENH: added new variant of MET_ValueToValue() with buffer size
MET_ValueToValue() didn't know the size of the `_toData` buffer, and so it was impossible to implement it safely, that is, with guaranteeing the buffer is not overwritten.

So , added a new variant where the buffer size is passed. Changed all but on callers to use the new variant. For the remaining caller, it wasn't clear what the buffer size is.

This allowed replacing the last sprintf in the codebase with snprintf, now that we can know the buffer size.