BUG: Do not allow user fields with the same names as keywords.
The MetaObject framework uses multiple keywords. Additionally it
allows the user to specify user fields (metadata). When the user
fields use the same names as the keywords, writing succeeds
but the resulting files are not valid. This commit adds a check so
that the user cannot name a user supplied field using a reserved
keyword.

resolves #54