ENH: API and file format version 2 - Position is not offset
Previously MetaIO considered origin and offset to have the same
meaning and transformMatrix and direction to have the same meaning.
However, in SpatialObjects, which are read and written as metaobjects,
position and offset have different meanings and so do transformMatrix
and direction.  So, prior to this commit, saving a scene graph of
spatialobjects could result in the loss of critical info.

ElementOrigin is introduced as the location of an object in its inherent
/ isolated / original space.  In particular, for images, it is the
"Image Position (patient)" tag which is the itk Origin() definition.

ElementDirection is introduced as the direction of acquisition for
the image. It follows the ITK definition of Direction().

Offset and transformMatrix (and centerofrotation) specify the
object-to-parent-object transform.  That is how an object (and all of
its children objects, aka. its scene subgraph) are transformed to map
into the parent object space.

Backward compatibility is preserved.   MetaIO defaults to using the
original API and the original file format.  If you wish to use the new
meanings (API) and write in the new file format, set the
APIVersion and FileFormatVersion variables to 2 (instead of 1 - default).