BUG: When writing a MetaIO file, the first line must be ObjectType
This rule was followed when writing all objects, except metaScene.  This
commit fixes metaScene.  Additionally, MET_Read/ReadType will look at
the first three tag/value pairs in a meta file to determine an object
type, so it is more robust, but it will throw warnings regarding
unknown fields if the first line isn't ObjectType.

Also reduces the specificity of the tests on ObjectName - allowing names
to appear as a substring anywhere in the value of the tag/value pair,
rather than requiring it to be at the beginning of the value.