ENH: Supress warnings when reading ObjectType and SubType
Reading objectType and objectSubType is done without knowledge of
the type, and thereby without knowing what fields will be present
or knowing their order of presentation.  Therefore, fields may be
encounted that are "unrecognized" when reading the Type / SubType
info.  Previously this would produce an "unrecognized field" warning.
This commit supressed those warnings.